Correcciones en la funcion de busqueda tras modificar la entidad de documento, funciones de creacion y eliminación de anotaciones

+++ WL.Application/Annotations/Commands/CreateAnnotationCommand.cs
@@ -4,7 +4,7 @@
     public long FromDocumentId { get; set; }
     public long ToDocumentTypeId { get; set; }
     public long ToEntityId { get; set; }
-    public int ToPublicationYear { get; set; }
+    public long ToPublicationYear { get; set; }
     public string ToNumber { get; set; }
     public long AnnotationTypeId { get; set; }
     public string Description { get; set; }
